export simplifiers in preparation for simplification
